refactor: move daemon-shared contracts out of commands
