+++ src/client-normalizers.ts
@@ -1,5 +1,5 @@
 import type { CommandFlags } from './core/dispatch.ts';
-import { screenshotFlagsFromOptions } from './commands/capture-screenshot-options.ts';
+import { screenshotFlagsFromOptions } from './contracts/screenshot.ts';
 import type { DaemonRequest, SessionRuntimeHints } from './daemon/types.ts';
 import { AppError } from './utils/errors.ts';
 import type { ScreenshotOverlayRef, SnapshotNode } from './utils/snapshot.ts';
