+++ src/commands/log-command-contract.ts
@@ -1,2 +1,1 @@
-export const LOG_ACTION_VALUES = ['path', 'start', 'stop', 'doctor', 'mark', 'clear'] as const;
-export type LogAction = (typeof LOG_ACTION_VALUES)[number];
+export * from '../contracts/logs.ts';
