+++ src/utils/parse-timeout.ts
@@ -1,0 +1,5 @@
+export function parseTimeout(value: string | undefined): number | null {
+  if (!value) return null;
+  const parsed = Number(value);
+  return Number.isFinite(parsed) ? parsed : null;
+}
